Add Camera-fix for custom-kernel
On late Android 4.4 (on some projects on 5.0) we updated the camerawrapper
and renamed our camera-blob to work properly.
In the past we used "vendor-camera.piranha.so", now we renamed it to
"camera.vendor.piranha.so".

On Galaxy Tab 2 we have 2 different camerablobs (depends on the kernel source)
and some custom kernel still use the old camera-blob-nameing while installing
the right camera-blob for the used kernel.

Create a symlink to use the new nameing of the camera-blob to get the camera
work properly instead FC-ing.